+++ CHANGES.md
@@ -50,6 +50,8 @@
 - `compute` method of `RegPull` now private
 - Renamed `data` to `fy` in `RegPull.evaluate` and `RegAsProxyPull.evaluate`
 - Made `get_layer` in `FullyConnected` private
+- Made `get_dense` and `load_embeds` in `CatEmbHead` private
+
 
 ## Depreciations